:iphone: design: 스타일 컴포넌트 구조 변경
