+++ client/src/components/header/Header.Style.tsx
@@ -2,6 +2,27 @@
 
 import { styled } from "styled-components";
 
+import { ReactComponent as Home } from "../../assets/images/icons/headerIcons/Home.svg";
+import { ReactComponent as Main } from "../../assets/images/icons/headerIcons/Main.svg";
+import { ReactComponent as Map } from "../../assets/images/icons/headerIcons/Map.svg";
+import { ReactComponent as CreateFeed } from "../../assets/images/icons/headerIcons/CreateFeed.svg";
+import { ReactComponent as Notifications } from "../../assets/images/icons/headerIcons/Notifications.svg";
+import { ReactComponent as NotificationsRedPoint } from "../../assets/images/icons/headerIcons/NotificationsRedpoint.svg";
+import { ReactComponent as Profile } from "../../assets/images/icons/headerIcons/Profile.svg";
+
+//assets style
+export const Logo = styled(Home)``;
+export const MainButtonStyle = styled(Main)``;
+export const MapButtonStyle = styled(Map)``;
+export const CreateFeedButtonStyle = styled(CreateFeed)``;
+export const NotificationsStyle = styled(Notifications)``;
+export const RedPointStyle = styled(NotificationsRedPoint)``;
+export const ProfileStyle = styled(Profile)`
+  padding: 15px;
+  background-color: #d7d7d7;
+  border-radius: 100px;
+`;
+
 export const HeaderContainer = styled.div`
   display: flex;
   flex-direction: row;
